Add FileLocation, rename location, position
Rename some types:

Location means where the "file" or file equivalent lives in the import
tree. For example, a file may import other files, and so on, in that
case each file is a "Location" for the checker/interpreter. Similarly, a
transaction, may import from a contract, which may import from another
contract. Each of these are "Location"s as well.

Position denotes a particular row/column *within* a file or file equivalent.