Change the project.AccountConfig type
Rename to just `Account`, given that it's in a config package and a
config struct, I think this is sufficient.
Store address/key as more useful types (rather than string). This avoids
the need to serialize/deserialize these fields explicitly all the time.
Instead, we just do the serialization when converting to JSON.