Implement "changelog" structure to support ledger trie
The changelog keeps track of the list of blocks at which each register
changed value. It has O(nlgn) inserts and O(lgn) searches.

The idea is to store the changelog and each register value for each block
at which it changes, in storage. The changelog can be cached in memory
for quicker lookups.