Change engine interface to just include Submit
This is the same interface as network.Engine, but without origin ID
(since it's always the same). I've re-added the error return because it
is useful for this particular use-case and I think it will be useful in
the future as well. The caller can always log the error if necessary, or
the engine can expose another method that logs errors if that behaviour
is desired.