Update mocks (again)
For some reason, my machine generates mock files with
```
import (
  x
  y
)
```

style of imports, but CI (and other dev machines) output with

```
import x
import y
```

style, causing the check-generated.sh script to fail.

This commit re-generates the mock files (this time in a Docker
container to ensure consistency with CI). This is a stop-gap until we
figure out where the inconsistency is coming from.