Correct prefix matching in badger iteration
Previous commit added ability to require a prefix in badger iteration.
It turns out that specifying a prefix in the badger iteration options in
insufficient to ensure that all results will have that prefix. You need
to use ValidForPrefix (rather than Valid) to detect when you've left the
prefix range and exit at that point.