move concurrency test to ingest engine
the network layer implements de-duplication, meaning that checking
duplication by checking what arrives at the consensus node doesn't work.
Even if the verifier double-verifies, the result approval is de-duped at
the network level. this change moves the logic of the concurrency test,
mostly unchanged to the ingest engine and checks for duplication at
entrance to the verifier engine.