update cluster state testss
* move payload integrity check to procedure
* add test for empty collection

For supporting empty collections, my thinking is: since the genesis
will always contain an empty collection, it is safe to omit storing the
collection for a block when it's empty, which is kind of the intuitive
thing to do anyway. The primary and indexing key will exist for the
empty collection by virtue of the genesis block.