review feedback
added enum to enumerate node info types
add PrivateKeys method and type to encapsulate both the private
information of the node intended for use in application logic.

Previously, where there were panics in the case that someone attempted
to retrive private data from a node for which only public info had been
specified, this has been converted to an error returned in either
`Private`, used for retrieving the private encoding of the node info, or
`PrivateKeys`, used for accessing the node's private keys within
bootstrapping logic.

ref: https://github.com/dapperlabs/flow-go/pull/3154#pullrequestreview-387400388