adds canonical range prefixes for iterating payload indexes
This change introduces payloadIterRange, which reliably returns start
and end prefixes for an iteration over a payload index. Previously,
these were computed ad-hoc in each function. Although they were computed
consistently, they had a subtle bug.

Previously the start prefix was code||boundary. This on its own would
omit all blocks at height boundary, since code||boundary is
lexicographically less than code||boundary||block_id, for any block ID.
However, when iterating in reverse, the `iterate` function adds a 0xff
byte to the start prefix, which ensures that the start prefix itself is
included in the iteration.

This extra byte has the side effect of expanding the prefix range of the
iteration to include all values of block_id that don't start with 0xff.
Since the probability of having a block ID beginning with 8 1's is 1/256,
on average, 1/256 blocks will be handled correctly when iterating through
their payloads.

This commit ADDs the range generation function and changes the iterate
function to add a 0x01 instead of 0xff, which makes this issue very
likely to appear in any given test run. It does not use the range
generation function, so this commit should still fail unit tests. The
next commit will incorporate the fix wherever it is needed.