benchmark for building a block
essentially tests the payload index iteration, as this is the critical
path for building a block for COL/CON nodes

result is that the payload check grows linearly as expected. Iterating
through one day's worth of blocks at 1 block/second takes on the order
of 500ms (half a second), so this check will quickly grow to dominate
the time taken by block proposal process.