update golangci-lint version
for some reason, GitHub is serving a 403 for requests for older release
assets of (at least this repo), including the version we use. you can't
even download it through their websit...

this must have happened recently because it breaks our CI process. this
change updates to the newest version, which is accessible.