only include cluster members in genesis QC signers
Previously the set of nodes that signed a cluster QC was filtered only
by whether they were a partner or not. It must also ensure only cluster
members sign the QC for their cluster.

This commit also removes a function that duplicates the cluster
generation mechanism, as this is now a public method of the protocol
package.