update sync engine to always respect block ID requests
currently, the sync engine will never double-request blocks, even if the
block is explicitly requested by the compliance layer. this means that
any transient error processing a block in the compliance layer will
result in a permanent halt to processing new blocks, since the
compliance layer will never receive the missing block it is requesting.

This updates the sync engine to always respect explicit requests for
block IDs, even if the block has already been received.