fix collector tests
The issue was that hotstuff was waiting a full minute before timing out
and moving to the next view. Since nodes don't all come up at the same
time, the first proposal consistently failed to propagate, then all
nodes would do nothing for a minute. With a test timeout of ~30s, this
caused the tests to fail. Even when extending the timeout, it still took
far too long for the nodes to agree on the view and be able to make
progress with the default settings. Changing the default and minimum
timeout to 1 second (from 60 seconds) addressed the issue.