update main readme
Move contributing section to CONTRIBUTING.md, following the example from
flow-go-sdk. I've kept it as-is for now, but it needs some love too --
lot's of TODOs, and would be good to make it consistent with the SDK doc
as well.

Simplified the main README a fair bit as a result of the above. Added
docs for common commands required during development and a note about
the local network.