add epoch manager engine
The epoch manager engine contains and manages the lifecycle for all
modules/engines that are epoch-dependent. This allows those
modules/engines to be unaware of epochs, and thus much simpler. This
design decision requires that channels used by these engines are also
epoch-specific, so the networking layer must support dynamic channels as
epochs come and go.