fix inconsistent encoding for epoch commit
This was a tricky bug. When you slice a byte array, the newly created
slice references the same backing memory as the original byte array. And
since the loop variable is re-used on each iteration, each new slice
ended up with the same backing memory. At the end of the loop, that
memory had the last node ID.