fix error equality checking in collector test
The test assumed that transaction properties were validated in a particular order,
which changed as part of implementing epochs. This commit changes the
test's assertion to be less specific (merely that an error occurred),
since the actual error that is returned is not directly related to the
failure condition we are testing, it is instead a side effect of the
failure condition we are testing.