add previous epoch to epoch status
Replace FirstBlock with PreviousEpoch in EpochStatus. Since we will
allow bootstrapping from arbitary blocks, we can no longer guarantee a
reference to the first block of an epoch (in particular the first) will
exist. On the other hand it remains necessary for the protocol state to
have a reference to the configuation to the previous epoch