Add emergency_sealed_blocks_total in compliance metrics
When the SealValidator encounters a seal which has at least one chunk with 0
signatures, and if required-verification-seal-approvals is 0, the validator
will assume that the seal was created in emergency mode, and will increment a
related counter in the compliance metrics. Note that the assumption that the
seal was created in emergency mode relies itself on the assumption that the
flag required-construction-seal-approvals is > 0.