modify EpochCommit to include only dkg keys
Previously the EpochCommit event contained both DKG keys and the
correspondence between those keys and node identifiers. This
correspondence is computable given both the EpochSetup and EpochCommit
events, but was duplicated here.

Since the DKG result key vector is opaque to the DKG smart contract, the
smart contract is unable to provide this correspondence in the service
event. Since it is unneeded, it is removed here.