Activate race-detection for those modules that are passing it
At present, the unit tests run as part of every CI build, but they do not run with with -race parameter which may be used to detect [data races](https://golang.org/doc/articles/race_detector).

Since data races may lead to instability and crashes, it's beneficial to ensure that we prevent data races from being introduced where possible. This runs unit tests with an explicit list of modules (rather than the recursive catch-all) and turns on `-race` for those that pass the build today, hoping others will join the list in the future.

To determine which modules "pass" with `-race` enabled, I've run `go test -race` in a loop and moved any directories that had a test failing because of a race condition move to the Makefile line (part of the unittest goal) without race detection.