Renamed to splitter and update multiplexer
Instead of having one multiplexer that is "channel aware", create a multiplexer that only consumes messages for one channel. Thereby, the multiplexer can forward all calls to the engines it wraps. Then it would be truly implementing the Engine interface and multiplex all calls to the wrapped engines.

In the multiplexer.network, we create one dedicated multiplexer.Engine per channel.