fix epoch contract deployment in fvm bootstrap
The model for clusters and qcs changed, which caused our manual
construction of them in Go to fail to be accepted when bootstrapping the
execution state. This commit removes unnecessary arguments entirely (QCs
and DKG pub keys) and simplifies the manual construction of clusters
(which are necessary) so that most of the heavy lifting is done by the
constructor in Cadence.