Retry unit-tests & integration-tests twice
- #988 introduced parallel testing on 1.15 / 1.16
- hence #988 multiplies the chance of failing under flaky tests by 2
- this should bring the probability back to where it was

Yes it is. But we're trying to merge fixes to flakiness (#987) and hotfixes (#1002), and the process of [requiring the branch to be up to date before merging](https://docs.github.com/en/github/administering-a-repository/defining-the-mergeability-of-pull-requests/managing-a-branch-protection-rule) means a branch can only merge if it's:
- right on top of master,
- the latest to be so,
- passing CI.

This PR attempts to reduce churn and manual work on the way to master.