[ci, crypto] Detect CPU arch and feed that info to blst
The Go bindings for BLST run with unconditional -D__ADX__ flag. While that is fine for newer
machines, GH Actions run tests on a variety of underlying architectures, some of which do not have
those CPU extensions. See:
https://docs.microsoft.com/en-ca/azure/virtual-machines/dv2-dsv2-series#dsv2-series

As a consequence, this runs detection before setting CGO compilation flags approaprately.