update payload write to use transaction for ER reads
Previously, when checking for ERs which were referenced by receipts in
the payload, we did not perform the read within the context of the
database transaction. This worked in the past, because blocks are
inserted sequentially, in separate database transactions. Now, we insert
a sequence of blocks (sealing segment) within the scope of one database
transaction when bootstrapping the state, so reads outside the
transaction context do not see pending ER writes, causing bootstrap
failure.