allow unverifiable block references below root height
In general the sealing segment may contain execution results which
reference blocks before the sealing segment. For nodes using the sealing
segment as an initial state, these blocks would be unknown.

It is safe to ignore these ERs, because they reference blocks prior to the
latest sealed block, by definition of the sealing segment. This commit
detects and ignores UnverifiablInput errors for these blocks.