modify logic for initializing spork beacon key
The logic attempted to initialize the spork beacon key based on whether
the node was a member of the identity table in its own root snapshot.
This assumed that nodes joining the network would not yet be a
participant when they join, which is typically true. However it is also
valid for nodes to use a later root snapshot, where they are already a
network member.

Instead, directly check whether the root snapshot is a spork root and
decide how/if to initialize the beacon key accordingly.