bootstrap with dynamic startup without writing snapshot
Rather than having the dynamic startup PreInit write a root snapshot, we
instead have it set a config field which is then used for bootstrapping.

We initialize Dynamic Startup only if:
* the protocol state is not bootstrapped, and
* there is no root snapshot file given

Upon completing Dynamic Startup, we store the root snapshot in the
NodeConfig.

When initializing the protocol state, we prioritize a root snapshot in
the NodeConfig and only read the root snapshot file if none is
configured.