Encode length of encoded payload value in 4 bytes
In v0, length of encoded payload value is encoded in 8 bytes.
In v1, length of encoded payload value is encoded in 4 bytes.

Payload decoding handles both v0 and v1.

Payload is encoded according to version received as parameter.
This allows gradual migration of other encoding formats using
payload, such as TrieUpdate and TrieProof.