handle registering node without machine account
Previously the epoch tests would pass in an argument representing a
machine account key, even for nodes which had no machine account. This
worked because the SDK accepted an empty account key and the argument
was ignored.

An SDK API change caused this to no longer work, because the SDK will
not accept empty account keys.

Instead, we handle the case of an empty machine account explicitly by
passing in a nil optional for this parameter.