Rename CheckpointQueue to TrieQueue
Rename CheckpointQueue to TrieQueue as suggested in PR review.

Since the name TrieQueue is not specific to checkpointing,
also moved it from complete package to wal package.

Putting it in the wal package makes it more clear TrieQueue
is specific to checkpointing and can prevent it from being
used for other purposes, getting modified, and introducing
side-effects to checkpointing.