Move Context fields into EnvironmentParams (Part 1)
I want to move facadeEnvironment into environment/ as well.  To do this, we need
to decouple facadeEnvironment from Context.  This is the first step.

Most parameters are logically be assoicated with a module; I've place those
parameters next to the module whenever possible.  Some Parameters, such as
Chain, are used by multiple modules; I'm keeping these in EnvironmentParams
for now.

Note that I'm migrating the parameters a few modules at a time rather than
wholesale to make the PRs more reviewable.