Refactor computer result collection
Refactor all execution result collection logic into ResultCollector.  In
particular, BlockCommitter and EventHasher are foled into the collector.

Note that BlockCommiter's Commit and EventHasher's Hash are now part of
CommitCollection and called inside executeCollection instead of executeBlock.
The order of operation got reordered a bit to improve pipeline efficiency.

Also note that the system chunk error logging got moved into executeTransaction
and is now logged on a per transaction level.  This is a no-op in reality since
the system chunk only has a single transaction.

gh: 3547