start pacemaker before processing pending blocks
We process pending blocks in the EventHandler prior to starting the main
event loop. This allows us to validate the pending blocks and process
their QC/TCs prior to accepting any new blocks. Part of this may involve
changing views, which will cause us to initiate a new timeout in the
TimeoutController.

The TimeoutController uses a context to gracefully shut down its ticker
goroutine, which is derived from the EventHandler's lifecycle context
and is set during the Start function. Currently, we process pending
blocks prior to starting the TimeoutController. So the TimeoutController context
will be nil at this point. If we process a pending block which causes us
to change views, then we will attempt to derive a child of a nil
context, which causes a panic!