rm node from approve-list when leaving
The core-contracts change modified the admin transaction to remove
a node to unstake it rather than deleting it (which was potentially
very dangerous as it would delete the stake as well).

This has the side effect that, for Access Nodes, since having 0 stake
and being allow-listed is enough to be a participant, refunding
stake does not actually cause removal.