Combine TransactionState & DerivedTransactionData into a single interface
TransactionState and DerivedTransactionData should eventually be validated and
committed together using 2PC.

The two are now folded into the storage.Transaction, which hides the
implementation detail and enables us to swap in a concurrency managed
implementation in the future.

This also moved away from using NewSimpleView in tests wherever possible.