skip typecheck lint on problematic files
When there are "real" lint errors, the linter will flood the output with
spurious errors from the typecheck linter (which is not a real linter:
https://github.com/golangci/golangci-lint/issues/2177).

typecheck errors are triggered when the Go compilation frontend fails -
it's not clear why a compilation error elsewhere triggers all the
spurious errors from AccessNodeBuilder etc. But disabling these makes it
easier to find the real comppilation error.