Add version, flag, and checksum to payload file
This commit also added --allow-partial-state-from-payload-file flag
to execution-state-extract program.

User needs to specify this flag allow partial payload file as input
(e.g. not all accounts) in order to prevent accidental use of
partial state during real migration.

The primary use case for payload file is for development, testing,
and debugging of migrations because it allows migrating a subset
of payloads instead of all payloads.