Add migration to fix refs to non-existent registers
This commit adds --fix-testnet-slabs-with-broken-references flag
to the migration program.

It uses a feature from onflow/atree#388 to fix 10 testnet registers
affecting 9 testnet accounts.

The 9 testnet accounts are hardcoded in this commit to prevent
accidentally applying this fix to any other accounts.

In testnet, broken references seem to have resulted from a bug
that was fixed 2 years ago by onflow/cadence#1565.

A broken reference is a `StorageID` referencing a non-existent register.
So far, only 10 registers in testnet (none on mainnet) were found to
contain broken references.