only validate root snapshot when loading file
Currently, we validate the root snapshot every time we startup, meaning
that for all startups after the first, we are validating the same thing
that we wrote to disk.

This causes a problem on v0.33, because some parts of the software
expect a sealing segment of length 590+, and others expect a sealing
segment of length 600+. In particular, this can cause a node to accept a
root snapshot file when bootstrapping, then reject the root snapshot
DB-backed object on subsequent startups, because the DB-backed snapshot
is returning an invalid sealing segment.