avoid reading the sealing segment in the scaffold
This change is to support nodes which bootstrapped with an invalid root
snapshot (too-short sealing segment) during the course of a v0.33
network. The software has been updated to again reject these invalid
snapshots for new nodes, but we need a temporary measure for nodes
which:
- bootstrapped using an invalid snapshot
- are unable or unwilling to re-bootstrap

In other words, we want a software version which will accomodate a
database that has been bootstrapped with an invalid snapshot (only for
the duration of v0.33 networks, this should not be ported to subsequent
major version!)

This PR:
- moves the setRootSnapshot call to after the database setup in both
  initial and subsequent startup cases
- changes the caching of root snapshot values to avoid reading the
  problematic root sealing segment

This should allow nodes with a bad root snapshot to boot, be a no-op for
nodes with a correct root snapshot, and also prevent nodes with a bad
root snapshot from producing further invalid snapshots, which would
propagate the underlying problem.