Make chunk assignment immutable
Add a new `AssignmentBuilder` type and move the `Add()` method to it.
`Assignment` now has no methods that mutate it. However, the initial
implementation doesn't do a deep copy of the data, so any changes to
the Builder will continue to be reflected in already-constructed `Assignment`s.

Some existing tests use other functions of the Assignment while still building it;
this behaviour hasn't been changed, and these tests continue to work as before.