Fix chunk Assigner TestByNodeID
Cleaned up the code for the test, and in the process discovered that the test
did not do what it claimed to:
Each verifier would be assigned a pair of adjacent chunks, and those pairs
would be overlapping; so the first chunk had one verifier, the next four had
two, the sixth had one, and the last four chunks were not assigned any verifiers.

Updated the test to use (and double-check) round-robin order of chunk assignment.
The specific ordering is not important, just that it is consistent
(and preferably easy to understand).