Convert chunk Assignment from map to slice, and enforce ordering
Per https://github.com/onflow/flow-go-internal/issues/5357 use a slice instead
of a map. This now enforces that chunks are assigned (added to the Builder)
in order of their Index (currently via a panic if the order is incorrect).

Necessitated changes to the helper for TestVerificationHappyPath, which was
not assigning some chunks (skipping odd-numbered chunks). Instead for now
we explicitly assign an empty verifier set.