Create TentativeEpoch interface
Updates EpochQuery.Next() to return a TentativeEpoch. (Will later be renamed
to NextUnsafe). All users of methods not present on TentativeEpoch now use
new method EpochQuery.NextCommitted(), which returns a regular Epoch
(to later be renamed to CommittedEpoch).

If NextUnsafe is called during the EpochCommitted phase, it returns an error.