Initial draft for Block.Payload refactor
To avoid duplicating the same information and creating potential malleability issues,
we decided to separate the Header's PayloadHash out from the rest of the fields.
As a 'compressed' representation of the block, the Header remains the same,
but Block no longer contains the PayloadHash, just the Payload.
Shared fields are moved from `Header` to `HeaderFields`.

This is a significant change (incomplete as of this commit).