Fix storage cache out of sync with db
Currently, in the new storage/store code, the memory cache
of some stores can have records that are no longer in the
database.

Specifically, the memory cache isn't updated when the database
record is deleted. So the memory cache can have records
that are not in the database.

This commit updates the following stores to remove the record
from the cache when the record is removed from database:
- commits.go
- events.go
- my_receipts.go
- transaction_results.go

This commit also adds tests to ensure cache and db are in sync
after records are removed.