Update DB.Reader() API to not return error
Currently, DB.Reader() returns error because MultiReader returns
error if provided reader list is empty.  However, this use case
is an implementation error and shouldn't happen.

This commit modifies NewMultiReader() to panic if the provided
reader list is empty.  It also updates DB.Reader() API to not
return an error because the only possible error was replaced
by a panic in this commit.