Update Reader.NewSeeker() API to not return error
Currently, Reader.NewSeeker() returns error because MultiSeeker
returns error if provided seeker list is empty.  However, this
use case is an implementation error and shouldn't happen.

This commit modifies NewMultiSeeker() to panic if the provided
seeker list is empty.  It also updates Reader.NewSeeker() API to
not return an error because the only possible error was replaced
by a panic in this commit.