Use fxamacker/golang-lru for faster del by prefix
Currently, deleting all keys that match a provided prefix
is too inefficient because hashicorp/golang-lru does
not have built-in support for that functionality.

The Remove() function in hashicorp/golang-lru cache
was used (as a stop-gap) to remove keys with matching
prefix by:
- getting all keys from cache
- comparing key prefix
- removing keys with matching prefix from cache

That was inefficient (as expected) because both getting
keys and removing keys are protected by a write lock.

This commit:
- uses fxamacker/golang-lru instead of hashicorp/golang-lru
- uses the faster RemoveFunc() instead of Remove()

Although this commit is an optimization, the focus was
on simplicity and maintainability (fast enough).

NOTE: fxamacker/golang-lru is a modified version of
hashicorp/golang-lru created today to add the faster
and more extensible RemoveFunc() to the simplelru as
an alternative to Remove(). I might open a PR at
hashicorp/golang-lru to add this feature.