Fix BadgerImpl Writer.Set & Delete args not safe to modify
Currently, docs for ReaderBatchWriter.Set() promises it is safe
to modify the key and value after Set() returns.  However, BadgerDB's
WriteBatch.Set() says users must not modify key and value until the
end of transaction.

This bug hasn't surfaced yet because of the way the functions are
currently used and leaving it as-is risks future problems.

This commit modifies ReaderBatchWriter.Set() & ReaderBatchWriter.Delete()
to allow the caller to modify the key and value as promised by our docs.