Add tests that modify argumments to Batch.Set(), etc.
This commit adds tests for storage Batch.Set() and Batch.Delete()
to ensure that arguments to these functions are safe to be modified
after functions return (before batch is committed).

This commit also adds test to show that BadgerDB's WriteBatch.Set()
argument are NOT safe to be modified before batch is committed.