Optimize iteration and seeking with BadgerDB
Currently, Iterator and Seeker causes BadgerDB to needlessly create
an iterator for every SSTables even for SSTables that don't contain
required data.

This commit sets IterationOptions.Prefix for Iterator and Seeker in
badgerimpl.  Prefix is used to narrow down the SSTables that iterator
picks up.  This can speed up creating and closing BadgerDB iterators
for databases that have many SSTables.