Add empty-migration flag to execution-state-extract
This commit adds empty-migration flag to execution-state-extract
cmd to estimate the minimum amount of time needed for state
migration.

For example, we can run this on m1-ultramem-96 with Skylake CPU
to get a baseline for spork migration downtime to decide if
we want to temporarily use a faster machine during spork, etc.