Create seq num payload if acct is created with keys
This commit creates sequence number payload(s) when adding account
public key(s) and sequence number is non-zero.

Normally sequence number is zero when keys are added to accounts,
and sequence number is incremented when key is used as proposal key.

However, it is possible that user provides account public key with
sequence number when accounts are created.