Create diff util for account public keys
This commit adds `diff-keys` command to the util program to compare
the account public keys before (v3) and after migration (v4)
and report any differences to a file (JSONL format).

If there are no differences found, then the output file is empty.

This command can be used after testnet and mainnet migrations to
verify the migrated account public keys.