Add transit script commands for cluster root block voting
The new commands are to be used by collector nodes.
Following creation of the cluster assignment by `cmd/bootstrap/run/clustering.go`,
Collector nodes retrieve the assignment, create a vote for the root block
of the collection cluster they are assigned to, and upload that vote.

Once enough votes have been received to construct QCs for all the cluster root
blocks, bootstrapping can continue to the next step (the consensus root block).